+++ src/domain/creator/decreaseFollowerCount.ts
@@ -1,6 +1,6 @@
 
-import retrieveData from './data/retrieve';
-import updateData from './data/update';
+import retrieveData from './repository/retrieve';
+import updateData from './repository/update';
 
 export default async function decreaseFollowerCount(creatorId: string): Promise<void>
 {
