+++ src/domain/creator/get.ts
@@ -1,5 +1,5 @@
 
-import retrieve from './data/retrieve';
+import retrieve from './repository/retrieve';
 import type CreatorView from './view/CreatorView';
 import createView from './view/createView';
