+++ src/domain/creator/getMe.ts
@@ -1,6 +1,6 @@
 
 import type Requester from '../authentication/Requester';
-import retrieve from './data/retrieve';
+import retrieve from './repository/retrieve';
 import type CreatorView from './view/CreatorView';
 import createView from './view/createView';
