+++ src/domain/creator/increaseFollowingCount.ts
@@ -1,6 +1,6 @@
 
-import retrieveData from './data/retrieve';
-import updateData from './data/update';
+import retrieveData from './repository/retrieve';
+import updateData from './repository/update';
 
 export default async function increaseFollowingCount(creatorId: string): Promise<number>
 {
