+++ src/domain/creator/repository/mapToUpdate.ts
@@ -1,0 +1,20 @@
+
+import type { RecordData } from '^/integrations/database/module';
+
+import type CreatorData from '../data/CreatorData';
+
+export default function mapToUpdate(data: CreatorData): RecordData
+{
+    return {
+        id: data.id,
+        fullName: data.fullName,
+        nickname: data.nickname,
+        email: data.email,
+        portraitId: data.portraitId,
+        // JoinedAt is not meant to be updated
+        popularity: data.popularity,
+        postCount: data.postCount,
+        followerCount: data.followerCount,
+        followingCount: data.followingCount
+    };
+}
