+++ src/domain/creator/repository/retrieve.ts
@@ -1,13 +1,13 @@
 
 import database from '^/integrations/database/module';
 
+import type CreatorData from '../data/CreatorData';
 import { RECORD_TYPE } from '../definitions/constants';
-import type CreatorData from './CreatorData';
-import mapRecord from './mapRecord';
+import mapFrom from './mapFrom';
 
 export default async function retrieve(id: string): Promise<CreatorData>
 {
     const record = await database.readRecord(RECORD_TYPE, id);
 
-    return mapRecord(record);
+    return mapFrom(record);
 }
