+++ src/domain/creator/repository/retrieveByStartNickname.ts
@@ -1,9 +1,9 @@
 
 import database, { RecordSort, SortDirections } from '^/integrations/database/module';
 
+import type CreatorData from '../data/CreatorData';
 import { RECORD_TYPE } from '../definitions/constants';
-import type CreatorData from './CreatorData';
-import mapRecord from './mapRecord';
+import mapRecord from './mapFrom';
 
 export default async function retrieveByStartNickname(nickname: string): Promise<CreatorData | undefined>
 {
