+++ src/domain/relation/explore.ts
@@ -1,7 +1,7 @@
 
 import type Requester from '../authentication/Requester';
-import retrieveCreatorsWithout from '../creator/data/retrieveWithout';
 import { SortFields as CreatorSortFields } from '../creator/definitions/SortFields';
+import retrieveCreatorsWithout from '../creator/repository/retrieveWithout';
 import createCreatorView from '../creator/view/createView';
 import retrieveDataByFollower from './data/retrieveByFollower';
 import SortOptions from './definitions/SortOptions';
